tleeop

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/ConstantsPP.java
@@ -3,9 +3,9 @@ package org.firstinspires.ftc.teamcode.drive;
 public class ConstantsPP {
 
     //DR4B values
-    public static double DR4B_GROUNDFLOORTURRETCLEARANCE = 0;
-    public static double DR4B_LOWJUNCTION = 185;
-    public static double DR4B_MIDHIGHJUNCTION = 350;
+    public static int DR4B_GROUNDFLOORTURRETCLEARANCE = 0;
+    public static int DR4B_LOWJUNCTION = 185;
+    public static int DR4B_MIDHIGHJUNCTION = 350;
 
     //V4B values
     public static double V4B_RETRACTED = 0.15;
